move transfer components to folder
